# openapi.yaml
openapi: 3.0.3
info:
  title: Task Management API
  version: 1.0.0
  description: >-
    REST API for an internal task management application supporting authentication,
    task lifecycle management, user administration, and dashboard summaries.
servers:
  - url: http://localhost:3000/api
    description: Local development server
security:
  - bearerAuth: []
...
paths:
...
  /settings:
    get:
      tags:
        - Settings
      summary: List Outlook integration settings
      description: Returns a paginated list of Outlook integration settings for the current workspace.
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/pageSize'
      responses:
        '200':
          description: Settings retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        outlookEmailAddress:
                          type: string
                        syncEnabled:
                          type: boolean
                        calendarId:
                          type: string
                        pollingIntervalMinutes:
                          type: integer
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      pageSize:
                        type: integer
                      totalItems:
                        type: integer
                      totalPages:
                        type: integer
        '400':
          description: Invalid pagination parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Settings
      summary: Create Outlook integration settings
      description: Stores a new Outlook integration payload for syncing calendar events and tasks.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - outlookEmailAddress
                - syncEnabled
              properties:
                outlookEmailAddress:
                  type: string
                  format: email
                syncEnabled:
                  type: boolean
                calendarId:
                  type: string
                pollingIntervalMinutes:
                  type: integer
                  minimum: 5
                  maximum: 1440
      responses:
        '201':
          description: Settings created successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  outlookEmailAddress:
                    type: string
                  syncEnabled:
                    type: boolean
                  calendarId:
                    type: string
                  pollingIntervalMinutes:
                    type: integer
        '400':
          description: Invalid request payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /settings/{settingId}:
    get:
      tags:
        - Settings
      summary: Get Outlook integration settings by ID
      description: Returns the Outlook integration configuration for a specific setting ID.
      parameters:
        - $ref: '#/components/parameters/settingId'
      responses:
        '200':
          description: Settings retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  outlookEmailAddress:
                    type: string
                  syncEnabled:
                    type: boolean
                  calendarId:
                    type: string
                  pollingIntervalMinutes:
                    type: integer
        '404':
          description: Settings not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags:
        - Settings
      summary: Update Outlook integration settings
      description: Applies a partial update to an existing Outlook integration setting.
      parameters:
        - $ref: '#/components/parameters/settingId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                outlookEmailAddress:
                  type: string
                  format: email
                syncEnabled:
                  type: boolean
                calendarId:
                  type: string
                pollingIntervalMinutes:
                  type: integer
                  minimum: 5
                  maximum: 1440
      responses:
        '200':
          description: Settings updated successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  outlookEmailAddress:
                    type: string
                  syncEnabled:
                    type: boolean
                  calendarId:
                    type: string
                  pollingIntervalMinutes:
                    type: integer
        '400':
          description: Invalid request payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Settings not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Settings
      summary: Delete Outlook integration settings
      description: Removes a setting record by ID.
      parameters:
        - $ref: '#/components/parameters/settingId'
      responses:
        '200':
          description: Settings deleted successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '404':
          description: Settings not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    page:
      name: page
      in: query
      description: Page number for pagination.
      schema:
        type: integer
        minimum: 1
        default: 1
    pageSize:
      name: pageSize
      in: query
      description: Number of records per page.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    settingId:
      name: settingId
      in: path
      required: true
      description: Unique identifier of the Outlook integration setting.
      schema:
        type: string
  schemas:
    ErrorResponse:
      type: object
      required:
        - code
        - message
        - timestamp
      properties:
        code:
          type: string
          example: INVALID_REQUEST
        message:
          type: string
          example: The request payload is invalid.
        timestamp:
          type: string
          format: date-time
          example: '2026-07-08T12:00:00Z'
...